+++ src-gui/src/renderer/components/other/ActionableMonospaceTextBox.tsx
@@ -18,6 +18,7 @@ type Props = {
   enableQrCode?: boolean;
   light?: boolean;
   spoilerText?: string;
+  truncate?: boolean;
 };
 
 function QRCodeModal({ open, onClose, content }: ModalProps) {
@@ -65,6 +66,7 @@ export default function ActionableMonospaceTextBox({
   enableQrCode = true,
   light = false,
   spoilerText,
+  truncate = false,
 }: Props) {
   const [copied, setCopied] = useState(false);
   const [qrCodeOpen, setQrCodeOpen] = useState(false);
@@ -101,6 +103,7 @@ export default function ActionableMonospaceTextBox({
           >
             <MonospaceTextBox
               light={light}
+              truncate={truncate}
               actions={
                 <>
                   {displayCopyIcon && (